Refactor settings dialog
